DM-13400: clean up and added working feedback when images take more that a 1/2 second.
